tun: Allow non-IFF_MULTI_QUEUE use of a IFF_MULTI_QUEUE device
Summary:
Currently if a persistent tun/tap device is configured with
IFF_MULTI_QUEUE all users must specify this flag, even if they only
want a single queue.

In AHV, we use a persistent tap device to add to the OVS switch before
asking libvirt to start the guest.  If we want to support optional
multiqueue this tap device needs to be multiqueue capable.  libvirt
however, only sets IFF_MULTI_QUEUE if more than one queue will be
used.

This causes problems if the user dynamically changes the number of
queues from N to 1 (or 1 to N). as the tap device will be in the wrong
mode.

Fix this by allowing IFF_MULTI_QUEUE to be omitted if it has already
been set.  This keeps the restriction on not allowing multiqueue on a
single queue device, but allows a single queue on a multiqueue device.

Note that we can't fix libvirt to set IFF_MULTI_QUEUE on single queue
devices, as a persistent device may not have this flag set.
Test Plan: dev
Target Release: obelix
Reviewers: felipe.franciosi, malcolm.crossley
Reviewed By: felipe.franciosi
CC: ahv
Differential Revision: http://codereview.nutanix.net/D13544